Make `utils` public and move `sorted_consts` into it.
