+++ hugr/src/lib.rs
@@ -149,7 +149,7 @@ pub mod macros;
 pub mod ops;
 pub mod std_extensions;
 pub mod types;
-mod utils;
+pub mod utils;
 
 pub use crate::core::{
     CircuitUnit, Direction, IncomingPort, Node, NodeIndex, OutgoingPort, Port, PortIndex, Wire,
